database/populate/requirements.txt
# Database and data processing
pandas>=1.3.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0

# ODS file support
openpyxl>=3.0.0
odfpy>=1.4.1  # Alternative ODS reader

# (Optional) For better performance
psycopg2-binary>=2.9.0  # PostgreSQL adapter
